Problems..
Started adding basic minion spawning code so that the player can summon friendly ai to fight with them. That part is fine, but unfinished. Trying to add this second type of ability has highlighted some flaws in the project structure and how code has been laid out so far. Next order of business is to restructure these parts of the ability system into its final form so that it is more expansible without becoming an unreadable mess.