Added the shield ability controller and first shield
Made a basic particle effect to represent a wind shield that uses the new shield class and is spanwed by the new shield spawner. It uses the health class that every damageable uses and interacts with all abilities. It has a positioning bug when it spawns it is spun 90 degrees. Also added an effect when ice spike collides.