Hacked in the shield and minion
Also, looked at the earth shatter particle to learn some about making complex particle systems. The shield spawns in front of the player camera direction + 5, this is hard coded and needs fixed. Cleaned up the minion spawning enough to have it in, though it is also hacked in. I did not add any AI to the minions yet, so you can spawn something at a location within a range, if you slap the minion script on it.