Added basic dash
Physics based dash, it does not feel good yet.
Also, decided to use my own ai. rv ai is a huge system that we won't need and would be quite a time sink to learn and lead to possible huge bugs if we don't learn it before putting it in the game.