Finished slow effect, and I suppose stun
set a slow amount to whatever fraction of 1 you want to reduce the target movement by (player or enemy). i.e. 80% slow = .8 slow amount. Currently  effect is not stackable or resettable...This could be changed. Started the Damage over Time effect but there is some problem with the coroutine. Should have this done tomorrow.