Fixed DoTs, Broke ability controller
DoTs need to have the coroutine started on the object that they're attached to.
Started refactoring the ability controller. It will be better designed code if the abilities get added to the player and have a key associated with them, then they can use the key to determine which type of controller they are using.